Update angularitics-splunk.js dynamic _sp
dynamically retrieving _sp by turning this into a function is necessary for anyone wanting to run setup in the module.config or not pull/compile the html of their app when running a testing framework like Karma. Since this code was being run prior to the initialization of the module as a dependency of your app, it will error with "define sp" in all consuming unit tests.