Add a timeout function around the data layer push for page views.
I ran into a problem when tracking virtual page views with route change success and dynamic page titles. Both the titles and the page views are triggered from route change success but the page view gets triggered before the page titles are set. In Google analytics the title from the previous page is recorded instead of the page title for the active route. Adding the timeout function solves this problem.